Cambios de la rama Prueba

+++ src/main/java/com/gym/monsterfit/entities/RutinaEjercicioEntity.java
@@ -1,8 +1,9 @@
 package com.gym.monsterfit.entities;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
+
+import java.util.ArrayList;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -31,7 +32,7 @@ public class RutinaEjercicioEntity {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @NotNull
